requirements.txt
Django==5.2.6
django-ckeditor==6.7.3
pillow==11.3.0
psycopg2-binary==2.9.11
python-decouple==3.8
djangorestframework==3.16.1